+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- support Swift Package Manager.
+
+### Fixed
+- fix Interface Builder render error.
+
 ## [1.9.6] - 2019-07-13
 
 ### Changed