+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.9.8] - 2019-08-30
 
 ### Fixed
 - when two thumbs overlap (`keepsDistanceBetweenThumbs == false`) let user drag lower thumb and not only upper thumb.